修改：     core/discrete_sample.hpp 	修改：     core/distribution.hpp 	修改：     core/find_peak.hpp 	修改：     core/graph.hpp 	修改：     core/particle_filter.hpp 	修改：     core/stochastic_node.hpp 	新文件：   helper/abstract_node_factory.hpp 	修改：     nodes/arithmetic_node.hpp 	修改：     nodes/const_node.hpp 	修改：     nodes/normal_node.hpp 	修改：     nodes/obs_bin_node.hpp 	修改：     nodes/obs_const_node.hpp 	修改：     nodes/phi_node.hpp 	修改：     nodes/uniform_node.hpp

+++ core/discrete_sample.hpp
@@ -20,7 +20,7 @@ namespace mcmc_utilities
       }
     std::vector<T_p> prob(cp.size());
     prob[0]=std::exp(pd.eval_log(cp[0]));
-    for(int i=1;i<cp.size();++i)
+    for(size_t i=1;i<cp.size();++i)
       {
 	prob[i]=prob[i-1]+std::exp(pd.eval_log(cp[i]));
       }
